+++ src/dataconnect/types.ts
@@ -134,16 +134,17 @@ export interface Generate {
 
 export interface JavascriptSDK {
   outputDir: string;
-  package?: string;
+  package: string;
   packageJSONDir?: string;
 }
+
 export interface SwiftSDK {
-  // Optional for Swift becasue XCode makes you import files.
-  outputDir?: string;
+  outputDir: string;
+  package: string;
 }
 export interface KotlinSDK {
   outputDir: string;
-  package?: string;
+  package: string;
 }
 
 export enum Platform {